Moves layout/topics-detail to layout/app/topics-detail
