+++ layout/app/topics-detail/topics-detail-content/component.js
