+++ layout/app/topics-detail/topics-detail-header/component.js
